requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
chromadb>=0.4.22
groq>=0.4.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
textblob>=0.17.1
python-dotenv>=1.0.0
pydantic>=2.5.0
newsapi-python>=0.2.7
ddgs>=1.0.0
scholarly>=1.7.11
arxiv>=2.1.0
pandas>=2.1.0
numpy>=1.24.0
aiohttp>=3.9.0
lxml>=4.9.0
html5lib>=1.1
python-dateutil>=2.8.0
feedparser>=6.0.10
xgboost>=2.0.3
scikit-learn>=1.3.2
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
yfinance>=0.2.40
lightgbm>=4.0.0
tensorflow>=2.13.0
keras>=2.13.0

openai>=1.0.0

# NEW: LangGraph dependencies for Phase 1
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
langchain>=1.0.0
langchain-core>=0.3.0

# =============================================
# NEW: Enterprise ESG Features (2026)
# =============================================

# ClimateBERT NLP Integration
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# SHAP/LIME Explainability
shap>=0.43.0
lime>=0.2.0.1

# Company Report Fetching & PDF Parsing
pdfplumber>=0.10.0
pypdf>=4.0.0
newspaper4k>=0.9.3.1

# Indian Data Sources
feedparser>=6.0.10  # Already included above, just noting

# Additional NLP
nltk>=3.8.0

# Camelot PDF Table Extraction
camelot-py[cv]>=0.11.0
opencv-python-headless>=4.8.0
PyMuPDF>=1.23.0

# Multi-Archive Cascade Retriever
requests-cache>=1.1.0
aiohttp>=3.9.0

# ESG Data APIs (WBA + WRI Aqueduct)
httpx>=0.27.0

# Company-Centric Knowledge Graph
neo4j>=5.23.0
langchain-experimental>=0.3.0
